Remove Theme Color option
When I was messing around with settings, I accidentally blew away all of my persistent config, and the theme color option was blank, which caused the web-UI to render in very strange ways.  We don't need that option, so I'm pulling it out for stability reasons.

This also removes other left over Web UI elements from the items we've pulled out that no longer are necessary.