Add support for token authentication with the provided DATA_URL (#5)
This is targeted to work with raw.githubusercontent.com for accessing
private repository data, with a generated Personal Access Token. That
Personal Access token is entered in as the "API_KEY" for the WagFam
settings.

Previously I had thought that these keys were only valid for up to 12
months. However, that is only true for "Fine Grained" tokens (which are
a new thing), but classic personal access tokens can be generated with
no expiration date. Woot! I still think we should likely implement the
key rolling feature, just in case we need it down the road after the
units ship to customers.

Thanks to this post: https://github.com/orgs/community/discussions/24744
for spelling out how to make this happen!